Updated visual styles for readability and UX improvements
This commit updates the styles of the display elements in adr.html and AdrMakerNext.kt to improve readability and user experience. Specifically, a glass-style has been added to various text elements to enhance their visibility against the background. Also, it modifies linear gradients to enhance the depth perception of the elements. This update aims to make the information on the page easier to read and digest and to make the UI more visually appealing. Additionally, the order of color stops in the linear gradients is rearranged; it starts from light at the top, getting darker towards the bottom for a more natural and eye-pleasing look. This change helps to create a more intuitive visual hierarchy among the elements.