Add Status class import and bgColorMap function
The commit added an import for the Status class in AdrMakerNext.kt. A new function, bgColorMap was introduced to map 'Adr' statuses to corresponding background colors. This function was then used in the svg() function to fill SVG elements related to 'Adr' status with the appropriate colors. This contributes to intuitive visual cues for different 'Adr' statuses in the UI.